add an async IIFE
Wrap the entire module in an Immediately Invoked Function Expression:

https://developer.mozilla.org/en-US/docs/Glossary/IIFE

But not just any IIFE. An *async* IIFE. This allows you to "await" asyncronous
values throughout the file.

This will be put to use in the next commit.

As a bonus, but not of much application here, you get a sort of namespacing
that comes with all IIFEs: things like `employeeArray` and `eventArray`, for
example, won't be accessible outside this file.