change Promises to async/await
This shows how the `async` function wrapper allows you to use async/await at a
more "top level".

This allows you to consicely create a couple of `const` values in one pass,
instead of creating a global object and then using a Promise to modify that
object.

There's an async "gotcha" here though that I'll address in the next commit.