move `await`s into a Promise.all
How it was before
=================

Two sequential `await` calls, which run one after the other.

What was wrong with this
========================

These two calls don't depend on each other, and there's no reason for them to
run sequentially/synchronously.

What changed
============

Put the two calls into a Promise.all.

Why this is better
==================

Promise.all resolves an array of Promises to an array of values, resolving the
Promises in parrallel. Now the two fetches can happen at the same time instead
of one after the other.